move new pure function into a method on los
